+++ src/test/java/sysc4806/group7/mini_shopify/MerchantTest.java
@@ -37,7 +37,6 @@ class MerchantTest {
     @Test
     void addProductToShop() {
         Product toaster = new Product("Toaster", "a toaster", 12.99);
-        Shop s1 = m1.createShop("Arthur's Appliances", "a shop");
         m1.addProductToShop(s1, toaster);
         assertTrue(s1.getProducts().contains(toaster));
     }
@@ -45,7 +44,6 @@ class MerchantTest {
     @Test
     void removeProductFromShop() {
         Product toaster = new Product("Toaster", "a toaster", 12.99);
-        Shop s1 = m1.createShop("Arthur's Appliances", "a shop");
         m1.addProductToShop(s1, toaster);
         assertTrue(s1.getProducts().contains(toaster));
         m1.removeProductFromShop(s1, toaster);
@@ -55,7 +53,6 @@ class MerchantTest {
     @Test
     void setProductQuantity() {
         Product toaster = new Product("Toaster", "a toaster",12.99);
-        Shop s1 = m1.createShop("Arthur's Appliances", "a shop");
         m1.addProductToShop(s1, toaster);
         m1.setProductQuantity(s1, toaster, 22);
         assertTrue(s1.getInventory().getItemQuantity(toaster) == 22);
